# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tangent Solutions - PHP Assessment",
        "description": "The auto-generated swagger documentation for the Tangent Solutions PHP Assessment.",
        "contact": {
            "name": "Chris Kempen",
            "email": "chris@phpalchemist.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/post/{post_id}/comment": {
            "get": {
                "tags": [
                    "Comment"
                ],
                "summary": "Retrieve a list of post comments",
                "description": "Retrieves a list of all of the post comments created on the system for a post",
                "operationId": "comment-index",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "description": "The ID of the post you want to retrieve comments for",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": "7"
                                                    },
                                                    "full_name": {
                                                        "type": "string",
                                                        "example": "Firstname Lastname"
                                                    },
                                                    "email_address": {
                                                        "type": "string",
                                                        "example": "email@domain.com"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "user_id": 1,
                                                    "full_name": "Firstname Lastname",
                                                    "email_address": "email@domain.com"
                                                },
                                                {
                                                    "user_id": 2,
                                                    "full_name": "Firstname Lastname",
                                                    "email_address": "email@domain.com"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Comment"
                ],
                "summary": "Add a new post comment",
                "description": "Add a new post comment to the system",
                "operationId": "comment-store",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "description": "The ID of the post you want to retrieve comments for",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": "1"
                        }
                    }
                ],
                "requestBody": {
                    "description": "New post particulars",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content",
                                    "user_id"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "This is an example of the content."
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user_id": {
                                            "type": "integer",
                                            "example": "12"
                                        },
                                        "full_name": {
                                            "type": "string",
                                            "example": "Firstname Lastname"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "email@address.com"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "full_name": {
                                                        "type": "array",
                                                        "items": {},
                                                        "example": [
                                                            "Invalid full name."
                                                        ]
                                                    },
                                                    "email": {
                                                        "type": "array",
                                                        "items": {},
                                                        "example": [
                                                            "The email has already been taken."
                                                        ]
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": {
                                                "full_name": [
                                                    "Invalid full name."
                                                ],
                                                "email": [
                                                    "The email has already been taken."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api_token": {
                "type": "apiKey",
                "description": "An API key associated with a user that's been added to the system.",
                "name": "api_key",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}